(TYP): Default to f64 type for floating point literals without postfix
This is documented at
https://doc.rust-lang.org/book/primitive-types.html

+++ src/main/kotlin/org/rust/lang/core/types/RustFloatType.kt
@@ -15,13 +15,13 @@ class RustFloatType(val kind: Kind) : RustPrimitiveTypeBase() {
 
         //
         // TODO(xxx):
-        //  The type of an unsuffixed integer literal is determined by type inference
-        //      > If an integer type can be uniquely determined from the surrounding program context, the unsuffixed integer literal has that type.
-        //      > If the program context under-constrains the type, it defaults to the 32-bit float
+        //  The type of an unsuffixed floating point literal is determined by type inference
+        //      > If an floating point type can be uniquely determined from the surrounding program context, the unsuffixed floating point literal has that type.
+        //      > If the program context under-constrains the type, it defaults to the 64-bit float
         //      > If the program context over-constrains the type, it is considered a static type error.
         //
         fun deduceUnsuffixed(@Suppress("UNUSED_PARAMETER") o: RustLitExprElement): RustFloatType =
-            RustFloatType(kind = RustFloatType.Kind.f32)
+            RustFloatType(kind = RustFloatType.Kind.f64)
     }
 
     enum class Kind { f32, f64 }
